Replace the ListMembers_FromTeamAtlassian_ShouldReturnManyMembers by ListMembers_FromFirstTeamOfLoggedUser_ShouldReturnManyMembers
listing the members of the team Atlassian is no more publicly accessible (probably an RGPD limitation)
To test the team members route we prefer to test on the members of the first team in which the logged user is.
This will ensure that we tests on a team on which we have some rights

Furthermore we have already the test GetTeams_FromLoggedUser_ShouldReturnManyTeams that require the logged user to be a member of at least one team, so it's a requirement that was already here.