Implement the OAuth2 authentication on BitBucket for LibGit2Sharp
Some code has been shared with the OAuth2 authentication on BitBucket for APIs
It's probably possible to share more code and improve OAuth2 support by implementing ReFreshToken behaviour, but it's not the goal right now.

So we will kept all that code internal until we have a good code coverage and be able to made significant improvements with the tests as a safe net.