Refactor the TeamsEndPoint to extract a TeamResource and clarify the distinction between operations that doesn't need to know the team name (like GetUserTeams) from the operations that have been moved in the TeamResource that need it.
Also move the GetPaginatedValues method that were in EndPoint class to a PaginatedListIterator internal class that allow to use the method anywhere we have access to the SharpBucketV2 instance.
Like that we can build Resource classes that represent a sub path of an EndPoint class without needing to duplicate the resource public methods as internal methods in the base EndPoint class.
To resume: external view is the same, but internally it's much simple.