Rework the EnumerateSearchCodeSearchResults signature to take a pageLen parameter instead of a max results parameter.
Since we return a lazy enumerable caller can use the Take(IEnumerable<T>,int) method from linq to specify the max number of results to return. So exposing and implementing a max parameter is a doubloon with the linq functionality.

The only advantage we get with our max parameter was a small optimization if the max is less than the default page len. To found back that capability we prefer to expose the pageLen parameter so the caller can do that optimization itself if he need it, and can even do it better than what we where doing.

note: an internal metric counter has been added in SharpBucket class so we can validate in tests that the requests are done lazily as expected. This is a temporarily solution while waiting progress on PR #71 to be able to work with mocks more easily in tests.