Fix exception handling in RequestExecutor
While coding the SearchCodeSearchResult POCO I've written an invalid POCO (arrays are not supported by the json deserializer that we use), and instead of getting a clear exception message, I get a BitBucketException including a valid response content message, which was really strange.

After having write few unit tests to reproduce the issue and read a bit the RestSharp code, I've take back json deserialization at our side to perform it only on successful requests, otherwise it was done even for HTTP Responses with code 400 or more, which was a none sense and cause exceptions which was stored in the IRestResponse result and that was polluting the exception handling.

I've also split the call to result.IsSuccessful in two tests to correctly make the distinction on issues "raised" by BitBucket from issues "raised" by RestSharp.

Note that we lost content negotiation for the deserialization in the process of taking it at our side. But since BitBucket send us back only JSON response when we have something to deserialize, it seems not required for now to duplicate the private code that select the right deserializer in function of the response content-type.