Add missing public Enumerate and EnumerateAsync overloads for ListMembers and ListFollowers in TeamResource
Breaking; By the way fix the return type of this methods which are users and not teams (since teams may follow it's not totally accurate but it's still better that saying it will return teams)

Also remove the obsolete methods in TeamsEndPoint to not have to also change their return types.
They are obsolete since V0.9.0, so it should be OK to remove that now.