Commit comments and PullRequest comments seems to be of the same nature
So instead of adding always more code in bloated classes like RepositoriesEndPoint and RepositoryResource
Create a new set of classes (CommitResource, CommentsResource and CommentResource) on a strict model where each resource class do itself the queries on its base URL and do not perform more query than that.

And Try to share the CommentsResource class between CommitResource and PullRequestResource.

Existing tests demonstrate that its OK when used with a PullRequestResource as parent resource.
Now, need to check if its OK with a CommitResource as parent resource, since PUT and DELETE verbs are not documented for the CommentResource in that case... Is that a miss in the documentation ? Or Sharing code like that is not what we should do ?