Add some tests that demonstrate that Get Put and Delete methods works for comments on commit like on comments on pull requests.
However analyze of the raw responses of BitBucket shows that Comment objects returned are not exactly the same between commits and pull requests.
Indeed, BitBucket return a commit_comment object with a commit reference in one case and a pullrequest_comment object with a pull requests reference in the other case.
But since having the id or the hash of the object that own this comment in the response is not really useful (you already known it since you provide it to perform the request), it seems ok to keep them using the same Comment POCO, and then the same Resource classes