Remove methods and properties that were obsolete since version 0.11.0 relative to GDPR.
Only the usernmae property in User object has been kept, with obsolete tag removed, and comment improved.
This is in fact still used for team users which may still appear in some query responses.
Even if with current deprecation of teams in favor of workspaces we may thing that those users may completely disappear, they are still there for now.